Material form used for new event creation
